t3200: tests for new branch.autosetupmerge option "simple"
In the previous commit a new autosetupmerge option was
introduced. Here the existing branch tests are extended
with three new cases testing this option - the obvious
matching-name and non-matching-name cases, and also a
non-matching-ref-type case.

The matching-name case needs to temporarily create
an independent repo to fetch from, as the general
strategy in these tests of using the local repo as
the remote precludes locally branching with the same
name as the "remote".

Signed-off-by: Tao Klerks <tao@klerks.biz>