reset: deprecate '--refresh', leaving only '--no-refresh'
The explicit '--refresh' option was needed in the past when '--quiet',
'reset.quiet', and/or 'reset.refresh' disabled the index refresh in 'reset
--mixed'. Those options have since either been deprecated or made to always
refresh the index by default, leaving only '--[no-]refresh' determining
whether the index is refreshed or not.

Because there is nothing other than '--no-refresh' to disable index refresh,
we do not need a '--refresh' option to counteract some other refresh
disabling.

To ensure users don't use what is effectively a no-op, remove '--refresh'
from the set of 'reset' options, as well as its usage in 'git stash'.

Signed-off-by: Victoria Dye <vdye@github.com>