Added Curl Option to Override Request Method v2
Added support for environment variable "CURLOPT_CUSTOMREQUEST"
and config option "http.customrequest" to allow setting the Curl
option to override the default request method used by HTTP Git
operations.  Primary reason for this is to allow support for
cloning repositories where only GET requests
are allowed by a local web proxy but not POSTs.  When cloning
a repo first a GET is made to the server and then a
POST is made to the "git-upload-pack" endpoint.  In some
corporate environments with strong controls
only GET requests are allowed to known repository hosts (such
as GitHub) through a web proxy to prevent data leakage.  Using this
new setting, a user can set the "CURLOPT_CUSTOMREQUEST=GET" env at runtime
or "http.customrequest = GET" in their config file which will
change the second request from a POST to a GET, bypassing
web proxy restrictions on the type of requests allowed.
Tested with GitHub, changing the request from POST to GET still
results in the expected behavior of the repo successfully being cloned.

This is v2 of this patch, which refactored the placement of the env
and added the ability to set the config file option.

Signed-off-by: agreenbhm <agreenbhm@gmail.com>