MyFirstContribution: also explain cover letter in GitGitGadget section
The "Sending Patches via GitGitGadget" section mentions that the PR
title and description will be used as the cover letter, but does not
explain what is a cover letter or what should be included in it.

Mention the purpose of the cover letter in that section, and give
examples for the title and description, leveraging the excerpt extracted
from the "Sending Patches with git send-email" section in the previous
commit.

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>